pi/requirements.txt
openwakeword>=0.6.0
pyaudio>=0.2.14
openai>=1.30.0
anthropic>=0.30.0
numpy>=1.24
python-dotenv>=1.0.0
